Updated back/up controls
Added an up control to SignUpActivity (LoginActivity is the parent of
SignUpActivity) and pressing up will lead you back to the parent.

Also added intent flags to control the back button so that we create a
new task (LoginActivity) but clear the old task (MainActivity) so that
pressing back on LoginActivity will simply exit the app.