Querying Parse
We will now query Parse to get up to 1000 usernames in ascending order.
This may need to be modified if we have more than 1000 users.

Also created new ParseConstants class that will store all of our keys.
This means that if we ever change the username class to
supercoolusernames, we can just change this here instead of everywhere
in our code.

Woohoo :smiley: :balloon: