Added options for Camera Icon
Now, clicking on the Camera Icon will bring up 4 options:
* Take picture
* Take video
* Choose picture
* Choose video

Stored in array ^

Also added a listener so that when one of these items are tapped on, the
dialog will be dismissed. No further action from this as of yet.

Also moved if statements to a switch for better performance.