User can now select a video from the Gallery
We will resolve the Content URI into the accepted URI and then check the
fileSize of the video. We will then (no matter what happens) close the
inputStream to avoid memory leaks and check to see if the fileSize is
above the limit. If it is, notify the user.
Will return at end of statements to make sure that we don't continue
onto the next line.