Added preparations for sending message
Will need a send() method.
Added createMessage() method that will define our message in a new class
called "Messages" (e.g: Before we had a class called User where we'd
store all of our users information).
In this new class, we will put keys in and tie them to their value.

Also added new getRecipientIds() method that will go through the whole
list adding those who are checked.

Currently, the message will not be sent as, the send() method has not
been created and is commented out in this commit to avoid errors.