Created Custom List Adapter
Remove old adapter that only showed username.
Added new adapter that will show name and also use the appropriate
layout to update the ImageView dependant on the media type.

Explanation:
MessageAdapter will take in the context and the messages array. These
will be stored in appropriate variables.
We will also run the super class in the context with the layout we're
adapting. We will then run getView() which will check whether the holder
exists already. If it doesn't, create the holder. If it does, get the
already created holder. This is important as, otherwise it leads to bad
performance if we're creating something that already exists which, will
cause lag whilst scrolling if we have a lot of messages.
Then, get the position of the message and check whether it is an
image/video and display the appropriate drawable. Finally, set the text
equal to the senders name and then return the converted view (The view
with all of the info updated).