Preparations for displaying our Image Message
Created ViewImageActivity.java activity and the layout
activity_view_image.xml to go with it.
Updated view portrait only.

InboxFragment updated with an onListItemClick() method. It will get the
message at the position, the file type and the url to the file on Parse.

If it is an image, we'll start up our ViewImageActivity and pass in the
Uri (As, at this point, we have converted the Url to a Uri) and then
start the Intent.