Scroll position is now maintained
Added an if statement to check if the adapter has been already created.
If it has, refill the adapter, if it hasn't, create it (This is more
efficient and also means we're not re-creating the list each time we
start the fragment).

When refilling the adapter, we'll get the adapter from the ListView and
then run refill() in the MessageAdapter.java file.
Here, we clear out our list and add in all of the new messages (We're
already grabbing our messages from Parse, the whole list of these
messages is being passed into refill()).
Finally, we'll our view to refresh itself because we have new data for
it.