Message will now be deleted from backend
After the image is displayed, we'll get our recipient ID's from our
message and store them in ids.

Then, we'll check if there is 1 recipient remaining. If there is, go
ahead and delete the message (In the background).
If there isn't, we've left the section to remove that one recipient.
This will be done on the next commit.