Will now remove the user if more than 1 recipients
Will remove current user from ids array, create a new array called
idsToRemove and then add the current user to that.
Finally, we'll use removeAll (Which is generally used to remove multiple
values) to remove an array with just the current user in from our
recipientIds column. We'll then save this change in the background.