Added Cancel Button
Set positioning values in activity_login.xml and removed from its style.
In activity_sign_up, Added new ImageView which will act as a divider
between the two buttons (Also added content description string for
this). Updated signupButton for positioning. Created new cancelButton
and also added positioning and a cancel_button_label for it.
Finally, updated SignUpActivity.java to store the cancel button and gave
it an onClick() method that will finish the activity once the button is
tapped on.