Updated layout of List Items
Image is now centred vertically, to the right.
Added a bit of padding (activity_horizontal_margin as defined in
dimens.xml).
Updated image to new purple image (This is only a placeholder image and
is actually changed by MessageAdapter.java).
Also updated TextView so that it is to the left of the messageIcon,
aligned to the left, has a placeholder text and has top and bottom
padding (message_vertical_padding as defined in dimens.xml).
Also updated text size to 16sp.

Removed divider (Set to null) in styles.xml. Makes no difference however
to current project as we already removed the dividers.

Updated MessageAdapter.java to use new, purple drawable images.