Added swipe to refresh functionality
Also removed unneeded support library.

In fragment_inbox.xml, the ListView is now surrounded by a
SwipeRefreshLayout.

Added 4 colors to colors.xml.

In InboxFragment.java, created a new SwipeRefreshLayout
(mSwipeRefreshLayout) with a refresh listener. Also set the colors from
colors.xml as the ColorSchemeResources.
Refactored our code to retrieve the messages in retreieveMessages().
Added if statement to check if refreshing, if refreshing, stop
refreshing.
Added onRefreshListener() so that retrieveMessages() will run when
refreshing.