Changes to GridView
Updated to have 3 columns now.
Created new user_item.xml with a RelativeLayout (White background) and
nameLabel (Empty text).

New UserAdapter.java is copy of MessageAdapter.java but with slight
modifications.

FriendsFragment.java checks if adapter exists. If it doesn't, creates it
and sets it. If it does, refills it.