RecipientsActivity is now a GridView
Extending Activity instead of ListActivity now.
New GridView variable.
Set attributes for GridView.
Set an empty view as the TextView for the GridView.
If statement to create or refill the adapter.

Added OnItemClickListener(). If more than 1 item is checked, the send
button becomes visible but, if not, it will be invisible.
We'll also display the visibility of the checkmark if the user is
selected or not as well.