Set up Push Notifications
Reformatted AndroidManifest.xml.
Added new permissions for push notifications.
Added new service and receiver tags for the PushService.

Added Receiver.java file for Parse-1.7.1 which requires a manual
receiver to deal with opening a Push Notification.
Here, the class deals with onPushOpen() which is ran from an intent
which is also passed in, creates a new intent with the same extras as
the one that ran onPushOpen() and then sets a flag to open a new task.
The intent opens MainActivity so, when the user taps on the Push
Notification, they will be brought to MainActivity.