Sending User-to-User Push Notifications
Added string where we set %s as the sender. The string shows "You have a
new Ribbit message from %s".

In RecipientsActivity.java, updated sendPushNotifications() method to
send the push notification. Here, we create a new ParsePush, set the
query to it and the message (Using getString() which takes 2 parameters,
one being set as %s) and then we send the push message in the
background.